requirements.txt
fastapi
uvicorn
azure-storage-blob
aiohttp
sqlalchemy
aiofiles
python-dotenv